mcp-server/requirements.txt
fastmcp>=2.0.0
tweepy>=4.14.0
python-dotenv>=1.0.0
